Temporal and AT TIME ZONE
Modify Example B so as to highlight that an argument for the "timezone" parameter can be provided as an expression other than a string literal.

Modified the "AS OF" example to demonstrate the value of using AT TIME ZONE to convert a local time zone value for the SYSTEM_TIME expression.
also added AT TIME ZONE for Next Steps (please check I did that correctly).