Updated Scale value : precision > 38
According to document The resulting scale is reduced to min(scale, 38 - (precision-scale)) but it is applicable when precision is grater than 38. So this would result in min(scale , (38-precision) + scale) and 38 - precession is always negative and 38 - (precision-scale) will be new scale . So Instead of  min(scale, 38 - (precision-scale)) 38 - (precision-scale) can be used.