adding "Force Strict Encryption"
SQL Server 2022 added "Force Strict Encryption" option, but it was missin from this page.
Also, since Strict Encryption require a certificate, it might be better to have step by step to enable this option, but I was not sure it is good place to add it here.

I also changed minor things, but I was not sure, do we still calls "Books Online"?

Update docs/tools/configuration-manager/protocols-for-mssqlserver-properties-flags-tab.md