Documentation of Azure AD group behavior in SQL DB
This documentation change is to address a few customer reported incidents in which the customer expected to be able to execute statements since they were logged in via an AAD group which had been granted those permissions. Instead, they were met with errors stating the user doesn't exist, because of the implicit user limitation on SQL DB. By calling out this limitation publicly, customers should be able to resolve this issue more quickly without having to engage customer support.