Add details around AAD Group users in the Implicit Schema and User Creation section
Updating the situations where a user can use a database without having a db user account to include Azure AD group membership. Expanding the note about implicit creation of an Azure AD user to include the error message when a user attempts to create or alter a schema (the error 2760 initially included applies to objects like tables, views, etc., but doesn't cover the scenarios for CREATE SCHEMA and ALTER SCHEMA).