DIsambiguate deprecation notice in disallow results from triggers
The `disallow results from triggers` configuration option documentation has a deprecation notice about the ability to return result sets from triggers. But a generic deprecation notice is used, with the wording "this feature".

It is unclear whether "this feature" refers to the configuration option or to the feature of returning result sets from triggers. With the sentence "We recommend that you set this value to 1." added after the notice, we can understand that the notice is about the feature, not the option. But still, that is confusing.

It appears older versions of the documentation, dating back to MSDN, were more clear:
> The ability to return result sets from triggers will be removed in a future version of SQL Server. Avoid returning result sets from triggers in new development work, and plan to modify applications that currently do this. To prevent triggers from returning result sets in SQL Server 2005, set the disallow results from triggers Option to 1. The default setting of this option will be 1 in a future version of SQL Server.

(Seen [here](https://stackoverflow.com/a/30915837/1178314).)

This change combines the older documentation with the generic deprecation notice.