Update execution-plans.md
clarify that the exception plan is not neccererily "the most efficient" but the best between the RP which were created. I came here as a result of a question in the forum. The current description confuse people to think that there is no reason to monitor executions and improve queries using hints for example, since the server already always create "the most efficient" EP.