Update sp-changedistributor-password-transact-sql.md
Added this:
If this is a remote distributor, then it needs to be run on all the Publisher Servers that is using this Distributor. If Distribution or publisher database is in AG, then it needs to be run on all the Distributor and Publisher nodes. It does not matter if the node is primary or secondary.