Update tablediff-utility.md
The document does not clarify where the file is and it is not clear that it is not exists if you do not add the replication feature of the SQL Server. There are many questions in the forum regarding where the file is and why it is not in the location people claim it should be.