Update aad-security-configure-tutorial.md
Added a note that describes limitations when an Azure AD user, that is part of the Azure AD group which is used as the Azure AD admin on the SQL Managed Instance, logs in to SQL Managed Instance. In this case, the `SUSER_SNAME` call returns the Azure AD user's name, but the `SUSER_SID` call returns `NULL`, since the given Azure AD user is not listed in the `sys.server_principals` table. Therefore, access to certain functionalities may be restricted.