Update collation-and-unicode-support.md
I've a number of changes in the table which details what is the default collation given the Windows locale. Some of these changes are based on actually testing running the SQL Server Setup with that locale. But there are also a couple which had the text "not available at server level" where I have made changes based on my results for other locals that had this. Thus, these changes needs to be verified. I have marked these changes with (*).

Permit me also point out that there are quite a few Windows locale entirely missing from the list.