Update add-sensitivity-classification-transact-sql.md
"SQL Server does not maintain a list of the mapping between LABEL and LABEL_ID.
Also, it seems most of the parameter of this T-SQL command are optional, e.g., when we create a classification using this T-SQL command, we can provide both LABEL and LABEL_ID. Or we can provide only one of them.   Also, we can set LABEL A with LABEL_ID 1 for object A, but also set LABEL A with LABEL_ID 2 for object B. 
if you want to ensure a one-to-one relationship between the LABEL and LABEL_ID, you need to specify both the LABEL and LABEL_ID correctly when you create a classification. "
My change is to trying to clarify the above.