Update assess.md
Need to add a note for CSS case deflection. Confirmed the following with Raj (PM) so adding a note to the prerequisite for our customers.

"If AMA already exists with proxy, SQL BPA just uses it. If AMA is not deployed then it installs and sets it up, at that point it doesn't set proxy. "

Please feel free to rephrase this message on the commit.