Update sql-machine-learning-services-windows-install-sql-2022.md
Further updates to installing python. The existing link to the downloads now only goes to a page where it says the files aren't available. The link that I've provided is to the same site but gets you to a page with the release that's required.
Also, earlier today, I suggested a change to the command for installing R. Similar changes are needed for Python when using named instances. The instructions as shown would not work. And it's not obvious how to put the name of the named instance into the commands. I find most people are trying ".\SQLEXPRESS" or "MACHINENAME\SQLEXPRESS", neither of which will work.