Remove References to AT TIME ZONE in FOR SYSTEM_TIME AS OF
The documentation states that `AT TIME ZONE` can be used in conjunction with `FOR SYSTEM_TIME AS FOR`. The example gives, however, gives an error:
> Incorrect syntax near 'AT'.

This is true if you use the clause against a `VIEW` or a `TABLE`, the same error occurs. [db<>fiddle](https://dbfiddle.uk/6Kv6-9o6)

It appears, therefore, that the documentation is incorrect, and the section should be removed, which this PR does.

*If wanted, however, I am happy to add an example that uses a variable of `datetimeoffset` with `AT TIME ZONE` used against it, but I would want to check out how this works before doing so the documentation is informative.*