Update alter-index-transact-sql.md
Better description of what STATISTICS_NORECOMPUTE actually does, also slightly reworded the warning for this option.

Update create-statistics-transact-sql.md

Rephrased the warning for the NORECOMPUTE option to be identical to the (proposed) warnings on the equivalent STATISTICS_NORECOMPUTE option on the ALTER INDEX page and on the ALTER TABLE index_option page.

Update alter-table-index-option-transact-sql.md

Better description of what STATISTICS_NORECOMPUTE actually does, also copied and reworded the warning that is provided for this option on the ALTER INDEX page.