Table Variables can be used in INSERT..EXEC statements
The documentation advised that Table variables can't be used in an `INSERT...EXEC` statement. This *might* have been true before ([Stack Overflow](https://stackoverflow.com/q/14622940/2029983) suggests it was back in SQL Server 2000), however, it's not true in SQL Server 2014+. I have removed that statement, as it not true.

I've also "tidied" the same statement to make it clear that it's the `INTO` clause a table variable can't be used in; `SELECT ... INTO... FROM @TableVariable` is acceptable syntax.