Update from-using-pivot-and-unpivot.md to add UNPIVOT syntax
I was looking for the UNPIVOT syntax, but the page only has PIVOT syntax. The most significant differences for UNPIVOT include no <aggregate function>, no<alias for the pivot table> and the pivot columns from the IN clause are not repeated in the SELECT clause. Also, there were "[" rather than "<" identifying placeholders to be replaced by column names.