Update hadr-windows-server-failover-cluster-overview.md
I felt the original text is a little confusing as we still have to use the VMs in the backend pool for the Load Balancer used for an AG Listener. 

I've tried to re-word to indicate that it should be any of the VMs that can be the primary replica of that AG, but not sure if there is another way of saying it?