update cmdlet from `Get-WmiObject` to `Get-CimInstance`
Get-WmiObject has been deprecated. using this cmdlet will give you an error in powershell core and powershell 7. Error below:

Get-WmiObject: The term 'Get-WmiObject' is not recognized as a name of a cmdlet, function, script file, or executable program.
Check the spelling of the name, or if a path was included, verify that the path is correct and try again.