Split requirements for inlining to make more understandable
The current list of requirements is a mixture of requirements for the function definition and requirements for the context it is being used in.
There are also duplicate requirements in the examples section. 

This change splits the list into requirements for the function AND another list for the requirements of the context. 
It also removes the duplicate "examples" and refers users to the concise list.