Update connectivity-settings.md
I am suggesting these modifications to ensure maximum clarity regarding the end of support for TLS 1.0 and 1.1.  

The section "### Upcoming retirement changes" already mentions the end of support; however, the section "## Minimum TLS version" is the header, and it deserves emphasis on the end of support, as well as the removal of any references to the previous TLS versions.