Update sys-fn-hadr-backup-is-preferred-replica-transact-sql.md
The behavior change on this function leads some issues on our customer's side in one support case. We found that the change was introduced in SQL 2019 CU20 and SQL 2022 CU2 to fix another case sensitive issue. The related fix info is missing in official document for both CUs. Customer requests for the official document on this change.