Update alerts-create.md
The field was wrongly mentioned as Reason Type instead of Resource Type. The description needs to be completed.

I'm not well-versed with this feature but came across this while preparing for DP-300 exam. If it is indeed Reason Type, please change the text above the table instead (where it is mentioned as Resource Type).